+++ src/statistics.h
@@ -5,7 +5,7 @@
 
 void compute_statistics(System *sys, Means means);
 
-void update_stations_measurements(System *sys, double delta);
+void update_stations_measurements(System *sys, long double delta);
 
 void update_mean_measures(Means *means, Station *stations, int run);
